requirements.txt
azure_storage==0.37.0
mysql_connector_repackaged==0.3.1
openai==0.25.0
python-dotenv==0.21.0
slack_bolt==1.16.0
